requirements.txt
boto3==1.28.3
pillow==10.3.0
python-ffmpeg==2.0.12